Avoid SElinux denial accessing IMEI information
Set persistent attribute to false so preferences are not
stored in shared preferences, this avoids the SElinux
denial:

  W/SharedPreferenc( 6950): type=1400 audit(0.0:4): avc: \
  denied { search } for name="com.android.settings" \
  dev="mmcblk0p25" ino=48929 scontext=u:r:radio:s0 \
  tcontext=u:object_r:system_app_data_file:s0 tclass=dir\
  op_res=-13 ppid=319 pcomm="main" tgid=1589\
  tgcomm="m.android.phone"

Observed when entering IMEI information in:
Settings > About Phone > Status > IMEI information.

It makes no sense to store this information in shared
preferences anyway so no functinoality is lost with this
change.

Change-Id: Id778bbadc8141167d679177b39e1d7a540d5752c
Signed-off-by: freak97 <lennart.1997@gmx.de>